refactor: meditation icon
