+++ front/src/components/atoms/embedding/MeditationIcon/MeditationIcon.jsx
